feat(mobile): add visibility toggle and
- Move `Person` model to his own class instead of inside the interface
- Updated `IPersonApiRepository` to support fetching people with an option to include hidden ones.
- Add icon to toggle visibility of hidden people in the People Collection page.
- Added functionality to hide/unhide a person from person detail page.
- Updated localization to include a new string for hidden people.
