+++ mobile/lib/pages/library/library.page.dart
@@ -224,7 +224,7 @@ class PeopleCollectionCard extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final people = ref.watch(getAllPeopleProvider);
+    final people = ref.watch(getAllPeopleProvider());
     return LayoutBuilder(
       builder: (context, constraints) {
         final isTablet = constraints.maxWidth > 600;
