+++ mobile/lib/pages/search/all_people.page.dart
@@ -13,7 +13,7 @@ class AllPeoplePage extends HookConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final curatedPeople = ref.watch(getAllPeopleProvider);
+    final curatedPeople = ref.watch(getAllPeopleProvider());
 
     return Scaffold(
       appBar: AppBar(
