+++ mobile/lib/widgets/search/search_filter/people_picker.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:immich_mobile/domain/models/person.dart';
 import 'package:immich_mobile/extensions/asyncvalue_extensions.dart';
 import 'package:immich_mobile/extensions/build_context_extensions.dart';
-import 'package:immich_mobile/interfaces/person_api.interface.dart';
 import 'package:immich_mobile/pages/common/large_leading_tile.dart';
 import 'package:immich_mobile/providers/search/people.provider.dart';
 import 'package:immich_mobile/services/api.service.dart';
@@ -22,7 +22,7 @@ class PeoplePicker extends HookConsumerWidget {
     final formFocus = useFocusNode();
     final imageSize = 60.0;
     final searchQuery = useState('');
-    final people = ref.watch(getAllPeopleProvider);
+    final people = ref.watch(getAllPeopleProvider());
     final headers = ApiService.getRequestHeaders();
     final selectedPeople = useState<Set<Person>>(filter ?? {});
